requirements.txt
# Core libs
pillow
psycopg2
django<1.10,>=1.9
django-cors-headers
djangorestframework<3.4
django-widget-tweaks==1.4.1
git+https://dev.izeni.net/izeni/django-accounts.git

# Dev tools
django-extensions
